fix: update field names and types in B-tree structures for consistency

+++ src/b_tree_leaf.rs
@@ -18,7 +18,7 @@ pub struct BTreeLeaf {
     layout: Layout,
     search_key: Constant,
     contents: BTreePage,
-    current_slot: usize,
+    current_slot: i32,
 }
 
 impl BTreeLeaf {
@@ -30,7 +30,7 @@ impl BTreeLeaf {
     ) -> BTreeLeaf {
         let contents = BTreePage::new(transaction.clone(), block_id, layout.clone());
 
-        let current_slot = contents.find_slot_before(search_key.clone()) as usize;
+        let current_slot = contents.find_slot_before(search_key.clone());
 
         BTreeLeaf {
             transaction,
@@ -48,11 +48,11 @@ impl BTreeLeaf {
     pub fn next(&mut self) -> bool {
         self.current_slot += 1;
 
-        if self.current_slot >= self.contents.get_number_of_records() as usize {
+        if self.current_slot >= self.contents.get_number_of_records() {
             self.try_overflow()
         } else if self
             .contents
-            .get_data_value(self.current_slot)
+            .get_data_value(self.current_slot as usize)
             .equals(self.search_key.value.clone())
         {
             true
@@ -62,14 +62,14 @@ impl BTreeLeaf {
     }
 
     pub fn get_data_record_id(&self) -> RecordID {
-        self.contents.get_data_record_id(self.current_slot)
+        self.contents.get_data_record_id(self.current_slot as usize)
     }
 
     pub fn delete(&mut self, record_id: RecordID) {
         while self.next() {
             let data_record_id = self.get_data_record_id();
             if data_record_id.equals(&record_id) {
-                self.contents.delete(self.current_slot);
+                self.contents.delete(self.current_slot as usize);
                 return;
             }
         }
@@ -94,8 +94,11 @@ impl BTreeLeaf {
             let new_block_id = self.contents.split(0, current_flag);
             self.current_slot = 0;
             self.contents.set_flag(-1);
-            self.contents
-                .insert_leaf(self.current_slot, self.search_key.clone(), record_id);
+            self.contents.insert_leaf(
+                self.current_slot as usize,
+                self.search_key.clone(),
+                record_id,
+            );
             let directory_entry = DirectoryEntry {
                 block_number: new_block_id.get_block_number(),
                 data_value: first_value,
@@ -104,8 +107,11 @@ impl BTreeLeaf {
         }
 
         self.current_slot += 1;
-        self.contents
-            .insert_leaf(self.current_slot, self.search_key.clone(), record_id);
+        self.contents.insert_leaf(
+            self.current_slot as usize,
+            self.search_key.clone(),
+            record_id,
+        );
 
         if !self.contents.is_full() {
             return None;
